Move package: com.ppdai.das.client.annotation
