+++ das-client/src/main/java/com/ppdai/das/client/transaction/DasAnnotationValidator.java
@@ -3,6 +3,7 @@ package com.ppdai.das.client.transaction;
 
 import java.lang.reflect.Method;
 
+import com.ppdai.das.client.annotation.DasTransactional;
 import org.springframework.beans.BeanInstantiationException;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanPostProcessor;
