+++ das-client/src/main/java/com/ppdai/das/client/transaction/DasTransactionalEnabler.java
@@ -4,6 +4,7 @@ package com.ppdai.das.client.transaction;
 import java.lang.reflect.Method;
 import java.util.Objects;
 
+import com.ppdai.das.client.annotation.DasTransactional;
 import com.ppdai.das.core.client.DalTransactionManager;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanDefinition;
