+++ das-client/src/test/java/com/ppdai/das/client/transaction/autowire/TransactionAnnoClass.java
@@ -1,7 +1,7 @@
 package com.ppdai.das.client.transaction.autowire;
 
 
-import com.ppdai.das.client.transaction.DasTransactional;
+import com.ppdai.das.client.annotation.DasTransactional;
 import org.springframework.stereotype.Component;
