+++ das-client/src/test/java/com/ppdai/das/client/transaction/javaConfig/TransactionConfig.java
@@ -1,7 +1,7 @@
 package com.ppdai.das.client.transaction.javaConfig;
 
 
-import com.ppdai.das.client.transaction.EnableDasTransaction;
+import com.ppdai.das.client.annotation.EnableDasTransaction;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
